name cleanup (#211)
* work

* adding some high-level documentation

* more rename

* clear output

* typo

* tiny

---------
