+++ docs/trulens_eval/api/tru_feedback.md
@@ -1,3 +1,3 @@
 # Feedback Functions
 
-::: trulens_eval.trulens_eval.tru_feedback
+::: trulens_eval.trulens_eval.feedback
