+++ mkdocs.yml
@@ -68,7 +68,7 @@ nav:
       - API Reference:
           - Tru: trulens_eval/api/tru.md
           - TruChain: trulens_eval/api/truchain.md
-          - Feedback Functions: trulens_eval/api/tru_feedback.md
+          - Feedback Functions: trulens_eval/api/feedback.md
   - Explain:
       - Installation: trulens_explain/install.md
       - Quickstart: trulens_explain/quickstart.md
