+++ trulens_eval/DEPRECATION.md
@@ -1,0 +1,16 @@
+# Deprecation Notes
+
+## Changes in 0.2.0; to be deprecated at 1.0.0
+
+- file tru_feedback.py renamed to feedback.py
+
+- file tru_db.py renamed to db.py
+    - class TruDB renamed to DB
+
+- file tru_app.py renamed to app.py
+    - class TruApp renamed to App
+
+- in file schema.py
+    - class App renamed to AppDefinition
+
+Old files and classes should function with a warning until 1.0.0 .
