+++ trulens_eval/trulens_eval/Leaderboard.py
@@ -8,8 +8,8 @@ from streamlit_extras.switch_page_button import switch_page
 st.runtime.legacy_caching.clear_cache()
 
 from trulens_eval import Tru
-from trulens_eval import tru_db
-from trulens_eval.tru_feedback import default_pass_fail_color_threshold
+from trulens_eval import db
+from trulens_eval.feedback import default_pass_fail_color_threshold
 from trulens_eval.ux import styles
 
 st.set_page_config(page_title="Leaderboard", layout="wide")
