+++ trulens_eval/trulens_eval/tru.py
@@ -1,10 +1,8 @@
-from datetime import datetime
 import logging
 from multiprocessing import Process
 import os
 from pathlib import Path
 import subprocess
-import sys
 import threading
 from threading import Thread
 from time import sleep
@@ -13,11 +11,11 @@ from typing import Iterable, List, Optional, Sequence, Union
 import pkg_resources
 
 from trulens_eval.schema import FeedbackResult
-from trulens_eval.schema import App
+from trulens_eval.schema import AppDefinition
 from trulens_eval.schema import Record
-from trulens_eval.tru_db import JSON
-from trulens_eval.tru_db import LocalSQLite
-from trulens_eval.tru_feedback import Feedback
+from trulens_eval.db import JSON
+from trulens_eval.db import LocalSQLite
+from trulens_eval.feedback import Feedback
 from trulens_eval.utils.notebook_utils import is_notebook, setup_widget_stdout_stderr
 from trulens_eval.util import SingletonPerName
 from trulens_eval.util import TP
@@ -107,7 +105,7 @@ class Tru(SingletonPerName):
         self,
         record: Record,
         feedback_functions: Sequence[Feedback],
-        app: Optional[App] = None,
+        app: Optional[AppDefinition] = None,
     ) -> Sequence[JSON]:
         """
         Run a collection of feedback functions and report their result.
@@ -156,7 +154,7 @@ class Tru(SingletonPerName):
 
         return list(evals)
 
-    def add_app(self, app: App) -> None:
+    def add_app(self, app: AppDefinition) -> None:
         """
         Add a app to the database.        
         """
@@ -219,7 +217,7 @@ class Tru(SingletonPerName):
                     "Evaluator is already running in this process."
                 )
 
-        from trulens_eval.tru_feedback import Feedback
+        from trulens_eval.feedback import Feedback
 
         if not fork:
             self.evaluator_stop = threading.Event()
